table_read/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "table_read"
version = "0.1.0"
description = "Screenplay -> Claude direction -> ElevenLabs voices -> table read MP3"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.92.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
tableread = "table_read.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
